feat: api client 클래스 구현
