+++ frontend/src/apis/clients/httpMethod.ts
@@ -1,0 +1,8 @@
+const HTTP_METHOD = {
+  get: 'GET',
+  post: 'POST',
+  patch: 'PATCH',
+  delete: 'DELETE',
+} as const;
+
+export default HTTP_METHOD;
